permissions check

+++ services/booking/ticket_test.go
@@ -10,6 +10,7 @@ import (
 
 	"team-project/database"
 	"team-project/services"
+	"team-project/services/authorization"
 	"team-project/services/booking"
 	"team-project/services/data"
 
@@ -41,6 +42,7 @@ type ListTicketTestCase struct {
 	mockedTicket  data.Ticket
 	mockedTickets []data.Ticket
 	mockedError   error
+	mockedAuthorize    bool
 }
 
 func TestValidateForm(t *testing.T) {
@@ -111,6 +113,7 @@ func TestGetAllTickets(t *testing.T) {
 			want:          http.StatusOK,
 			mockedTickets: []data.Ticket{},
 			mockedError:   nil,
+			mockedAuthorize:    true,
 		},
 		{
 			name:          "Get_Tickets_500",
@@ -118,21 +121,32 @@ func TestGetAllTickets(t *testing.T) {
 			want:          http.StatusInternalServerError,
 			mockedTickets: []data.Ticket{},
 			mockedError:   errors.New("db error"),
+			mockedAuthorize:    true,
+		},
+		{
+			name:          "Get_Tickets_403",
+			url:           "/api/v1/tickets",
+			want:          http.StatusForbidden,
+			mockedAuthorize:    false,
 		},
 	}
 
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
+	defer func(){authorization.AdminRole = authorization.CheckAccess}()
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			ticketMock := database.NewMockTicketRepository(mockCtrl)
-			ticketMock.EXPECT().AllTickets().Return(tc.mockedTickets, tc.mockedError)
-			database.TicketRepo = ticketMock
-
+			authorization.AdminRole = func(w http.ResponseWriter, r *http.Request) bool{
+				return tc.mockedAuthorize
+			}
+			if tc.mockedAuthorize {
+				ticketMock := database.NewMockTicketRepository(mockCtrl)
+				ticketMock.EXPECT().AllTickets().Return(tc.mockedTickets, tc.mockedError)
+				database.TicketRepo = ticketMock
+			}
 			rw := httptest.NewRecorder()
 			req, _ := http.NewRequest(http.MethodGet, tc.url, nil)
-			//http.HandlerFunc(.GetAllTickets).ServeHTTP(rw, req)
 			router.ServeHTTP(rw, req)
 
 			if rw.Code != tc.want {
@@ -151,6 +165,7 @@ func TestGetOneTicket(t *testing.T) {
 			want:         http.StatusOK,
 			mockedTicket: testData,
 			mockedError:  nil,
+			mockedAuthorize:   true,
 		},
 		{
 			name:         "Get_Tickets_500",
@@ -159,15 +174,26 @@ func TestGetOneTicket(t *testing.T) {
 			want:         http.StatusInternalServerError,
 			mockedTicket: data.Ticket{},
 			mockedError:  errors.New("db error"),
+			mockedAuthorize:   true,
+		},
+		{
+			name:          "Get_Tickets_403",
+			url:           "/api/v1/ticket/fcb33af4-40a3-4c82-afb1-218731052309",
+			want:          http.StatusForbidden,
+			mockedAuthorize:    false,
 		},
 	}
 
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
+	defer func(){authorization.LoggedIn = authorization.CheckAccess}()
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			if tc.name != "Get_Tickets_400" {
+			authorization.LoggedIn = func(w http.ResponseWriter, r *http.Request) bool{
+				return tc.mockedAuthorize
+			}
+			if tc.mockedAuthorize{
 				ticketMock := database.NewMockTicketRepository(mockCtrl)
 				ticketMock.EXPECT().GetTicket(tc.id).Return(tc.mockedTicket,
 					tc.mockedError)
@@ -176,7 +202,6 @@ func TestGetOneTicket(t *testing.T) {
 
 			rw := httptest.NewRecorder()
 			req, _ := http.NewRequest(http.MethodGet, tc.url, nil)
-			//http.HandlerFunc(GetOneTicket).ServeHTTP(rw, req)
 			router.ServeHTTP(rw, req)
 			if rw.Code != tc.want {
 				t.Errorf("Expected: %d , got %d", tc.want, rw.Code)
@@ -202,6 +227,7 @@ func TestCreateTicket(t *testing.T) {
 			want:         http.StatusCreated,
 			mockedTicket: ticket,
 			mockedError:  nil,
+			mockedAuthorize:    true,
 		},
 		{
 			name:         "Post_Tickets_500",
@@ -209,6 +235,7 @@ func TestCreateTicket(t *testing.T) {
 			want:         http.StatusInternalServerError,
 			mockedTicket: ticket,
 			mockedError:  errors.New("db error"),
+			mockedAuthorize:    true,
 		},
 		{
 			name:         "Post_Tickets_406",
@@ -216,15 +243,26 @@ func TestCreateTicket(t *testing.T) {
 			want:         http.StatusNotAcceptable,
 			mockedTicket: data.Ticket{},
 			mockedError:  errors.New("validation failure"),
+			mockedAuthorize:    true,
+		},
+		{
+			name:          "Get_Tickets_403",
+			url:           "/api/v1/ticket",
+			want:          http.StatusForbidden,
+			mockedAuthorize:    false,
 		},
 	}
 
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
+	defer func(){authorization.AdminRole = authorization.CheckAccess}()
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			if tc.name != "Post_Tickets_406" {
+			authorization.AdminRole = func(w http.ResponseWriter, r *http.Request) bool{
+				return tc.mockedAuthorize
+			}
+			if tc.name != "Post_Tickets_406" && tc.mockedAuthorize {
 				ticketMock := database.NewMockTicketRepository(mockCtrl)
 				ticketMock.EXPECT().CreateTicket(tc.mockedTicket).Return(tc.mockedError)
 				database.TicketRepo = ticketMock
@@ -262,6 +300,7 @@ func TestUpdateTicket(t *testing.T) {
 			want:         http.StatusOK,
 			mockedTicket: ticket,
 			mockedError:  nil,
+			mockedAuthorize:    true,
 		},
 		{
 			name:         "Patch_Tickets_500",
@@ -269,6 +308,7 @@ func TestUpdateTicket(t *testing.T) {
 			want:         http.StatusInternalServerError,
 			mockedTicket: ticket,
 			mockedError:  errors.New("db error"),
+			mockedAuthorize:    true,
 		},
 		{
 			name:         "Patch_Tickets_406",
@@ -276,15 +316,26 @@ func TestUpdateTicket(t *testing.T) {
 			want:         http.StatusNotAcceptable,
 			mockedTicket: data.Ticket{},
 			mockedError:  errors.New("validation failure"),
+			mockedAuthorize:    true,
+		},
+		{
+			name:          "Get_Tickets_403",
+			url:           "/api/v1/ticket/fcb33af4-40a3-4c82-afb1-218731052309",
+			want:          http.StatusForbidden,
+			mockedAuthorize:    false,
 		},
 	}
 
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
+	defer func(){authorization.AdminRole = authorization.CheckAccess}()
 
 	for _, tc := range tests {
+		authorization.AdminRole = func(w http.ResponseWriter, r *http.Request) bool{
+			return tc.mockedAuthorize
+		}
 		t.Run(tc.name, func(t *testing.T) {
-			if tc.name != "Patch_Tickets_406" {
+			if tc.name != "Patch_Tickets_406" && tc.mockedAuthorize {
 				ticketMock := database.NewMockTicketRepository(mockCtrl)
 				ticketMock.EXPECT().UpdateTicket(tc.mockedTicket).Return(tc.mockedError)
 				database.TicketRepo = ticketMock
@@ -307,11 +358,12 @@ func TestUpdateTicket(t *testing.T) {
 func TestDeleteTicket(t *testing.T) {
 	tests := []ListTicketTestCase{
 		{
-			name:        "Delete_Ticket_404",
+			name:        "Delete_Ticket_200",
 			id:          uuid.Must(uuid.Parse("fcb33af4-40a3-4c82-afb1-218731052309")),
 			url:         "/api/v1/ticket/fcb33af4-40a3-4c82-afb1-218731052309",
-			want:        http.StatusNotFound,
+			want:        http.StatusOK,
 			mockedError: nil,
+			mockedAuthorize:    true,
 		},
 		{
 			name:        "Delete_Tickets_500",
@@ -319,19 +371,30 @@ func TestDeleteTicket(t *testing.T) {
 			url:         "/api/v1/ticket/fcb33af4-40a3-4c82-afb1-218731052309",
 			want:        http.StatusInternalServerError,
 			mockedError: errors.New("db error"),
+			mockedAuthorize:    true,
+		},
+		{
+			name:          "Get_Tickets_403",
+			url:           "/api/v1/ticket/fcb33af4-40a3-4c82-afb1-218731052309",
+			want:          http.StatusForbidden,
+			mockedAuthorize:    false,
 		},
 	}
 
 	mockCtrl := gomock.NewController(t)
 	defer mockCtrl.Finish()
+	defer func(){authorization.AdminRole = authorization.CheckAccess}()
 
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-
-			ticketMock := database.NewMockTicketRepository(mockCtrl)
-			ticketMock.EXPECT().DeleteTicket(tc.id).Return(tc.mockedError)
-			database.TicketRepo = ticketMock
-
+			authorization.AdminRole = func(w http.ResponseWriter, r *http.Request) bool{
+				return tc.mockedAuthorize
+			}
+			if tc.mockedAuthorize {
+				ticketMock := database.NewMockTicketRepository(mockCtrl)
+				ticketMock.EXPECT().DeleteTicket(tc.id).Return(tc.mockedError)
+				database.TicketRepo = ticketMock
+			}
 			rw := httptest.NewRecorder()
 			req, _ := http.NewRequest(http.MethodDelete, tc.url, nil)
 			router.ServeHTTP(rw, req)
